+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.1.8+build.32] - 2026-02-05
+
+### Build
+
+Updating interfaces to not depend on ros2 but on vyra interfaces. Therefor we can depict not only ros2 as interface but all transport protocols defined in com/transport
+
+
+## [0.1.8+build.31] - 2026-02-05
+
+### Build
+
+Bugfixing: redis->callable->self.redis_client
+
+
 
 ## [0.1.0] - 2025-06-11